Create button strip out of Job Actions (#26)
TLDR;
The behavior around running and enabling/disabling script runner scheduled jobs was confusing and split into 2 button groups.

We've pulled the Run and Enable/Disable buttons down into the button group with the Save button. A bug was fixed with enable/disable so it will save the schedule once it's clicked and no longer requires clicking save. The Run button has become a Save and Run button.